requirements.txt
playsound==1.3.0
python-telegram-bot==13.10
PyYAML==6.0
selenium==3.141.0
webdriver_manager==3.2.2